First attempt at adding particles
In attempting to account for future scenarios in which multiple rates for the different stages of the cell-particle interaction cycle can be input, I have not managed to model the desired behaviour properly. Splitting of internalised particles between cells when a cell proliferates is also a mechanism that I haven't yet input.