I fixed one bug and created another
By letting the amount of particles hovering over a matrix site be a fraction of the total number of free particles over the total number of matrix sites, it is likely (and probable) that free_prtcls will be a fraction. I thought initially that this would be fine - "it's just an average! just it in the rate of a poisson distribution and you'll still get a whole number". But it has proven to create more bugs and so I will try an ulterior method of distributing the particles amongst the grid sites at the beginning of each timestep.