Fixed bugs and distinct particles per cell
The method used in this version is to randomly distribute (as evenly as possible though) the particles across all sites in the petri dish prior to each timestep and not allow for fractions of a particle to be included. The issue here is that sometimes, the number of particles to interact with x which is derived from a poisson dristribution is greater than the number of particles available to the cell in the matrix site. I have solved this by using a simple max() function, but I'll look at solving this another way another time - perhaps by letting this occur while drawing particles from neighbouring cells