Fixed proliferation and visualisation with particles
This model incorporates the binomial inheritance of interacting/internalised nanoparticles when a parent cell splits into 2 daughter cells as it is described in the paper "Asymmetry of nanoparticle inheritance upon cell division: Effect on the coefficient of variation", Lijster and Aberg. It also keeps a tally of the number of free, interacting and internalised particles per timestep in the system (all the more useful for assessing internalised Vs associated nanoparticles) and implements a new visualisation of the petri dish. In this simulation, marker size indicates the cell proliferation cycle phase a cell is in, cell opaqueness indicates how many particles a cell has internalised, and cell outline width indicates how many particles a cell is interacting with. The simulation is yet to allow the input of different cell-particle interaction transition probabilities per cell proliferation cycle phase.