Fixed the internalisation model
For each cell, the number of particles in stage 0 of the model for each cell is drawn from bin(free particles on site, rate_interacts). For each cell, the number of particles that transition from stage i of the model to stage i+1 is drawn from bin(particles in stage i, (1-num_internalised/max_prtcl)*rate_i_to_iplus1). If the number of particles that would transition exceeds the carrying capacity, an 'if' statement caps it.