Fixing the particles in each stage
Due to wanting to keep the total number of particles allowed to transition from a stage in one timestep static in spite of successful transitions on the same timestep, save the number of successes without yet updating the cell_prtcl array. This better reflects the hypoexponential distribution of waiting times that the system experiences.